+++ tbot/log.py
@@ -185,3 +185,12 @@ def message(
     :param Verbosity verbosity: Message verbosity
     """
     return EventIO(["msg", str(verbosity)], msg, verbosity=verbosity, text=str(msg))
+
+
+def warning(msg: typing.Union[str, c]) -> EventIO:
+    """
+    Emit a warning message.
+
+    :param str msg: The message
+    """
+    return message(c("Warning").yellow.bold + ": " + msg, Verbosity.QUIET)